Datapoint init script: Allow a user-uploaded buildstock.csv to be used if provided. If not provided, sampling script is run.

+++ project_resstock_dsgrid/scripts/worker_initialization/datapoint_init.sh
@@ -21,6 +21,7 @@ if ! [ -f $FILENAME ]; then
   CNT="0"
   
   # Download and extract weather files
+  echo "Retrieving weather files."
   while [ $NUMEPWS -le "1" ]; do
   
     `curl --retry 10 -O "$1"`
@@ -45,14 +46,24 @@ if ! [ -f $FILENAME ]; then
   
   cd ..
 
-  # Run sampling script
-  NUMDATAPOINTS=`awk -F\"maximum\": 'NF>=2 {print $2}' analysis.json | sed 's/,//g' | head -n1 | xargs` # Yes, this is gross.
-  echo "NUMDATAPOINTS is $NUMDATAPOINTS"
-
+  # Run sampling script; if script has been uploaded, use that instead.
   OUTCSV="buildstock.csv"
-  `ruby lib/resources/run_sampling.rb -p NA -n $NUMDATAPOINTS -o $OUTCSV`
+  if ! [ -f "lib/housing_characteristics/$OUTCSV" ]; then
+  
+    echo "Generating buildstock.csv sampling results."
+    
+    NUMDATAPOINTS=`awk -F\"maximum\": 'NF>=2 {print $2}' analysis.json | sed 's/,//g' | head -n1 | xargs` # Yes, this is gross.
+    echo "NUMDATAPOINTS is $NUMDATAPOINTS"
 
-  cp "lib/resources/$OUTCSV" "lib/housing_characteristics/$OUTCSV"
+    `ruby lib/resources/run_sampling.rb -p NA -n $NUMDATAPOINTS -o $OUTCSV`
+
+    cp "lib/resources/$OUTCSV" "lib/housing_characteristics/$OUTCSV"
+    
+  else
+  
+    echo "Using uploaded buildstock.csv."
+  
+  fi
   
 else
